requirements.txt
networkx
fastapi
uvicorn
httpx
pytest-asyncio